fix macos apple clang 15

+++ src/ContainerCarti.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 
 std::ostream& operator<<(std::ostream& os, const ContainerCarti& container) {
-    for (int i = 0; i <= container.carti->size() - 1; i++) {
+    for (size_t i = 0; i <= container.carti->size() - 1; i++) {
         os << (*container.carti)[i] << " ";
     }
     return os;
@@ -15,7 +15,7 @@ std::vector<Carte>* ContainerCarti::getContainer() const {
 ContainerCarti::ContainerCarti(const ContainerCarti &other): carti(new std::vector<Carte>) {
     if (other.carti != nullptr) {
         carti->reserve(other.carti->size());
-        for (int i = 0; i <= other.carti->size() - 1; i++) {
+        for (size_t i = 0; i <= other.carti->size() - 1; i++) {
             carti->push_back((*other.carti)[i]);
         }
     }
@@ -28,7 +28,7 @@ ContainerCarti& ContainerCarti::operator=(const ContainerCarti &other) {
         if (other.carti != nullptr) {
             carti->clear();
             carti->reserve(other.carti->size());
-            for (int i = 0; i <= other.carti->size() - 1; i++) {
+            for (size_t i = 0; i <= other.carti->size() - 1; i++) {
                 carti->push_back((*other.carti)[i]);
             }
         }
@@ -38,7 +38,7 @@ ContainerCarti& ContainerCarti::operator=(const ContainerCarti &other) {
 
 ContainerCarti& ContainerCarti::operator+=(const ContainerCarti &other) {
     if (other.carti != nullptr) {
-        for (int i = 0; i <= other.carti->size() - 1; i++) {
+        for (size_t i = 0; i <= other.carti->size() - 1; i++) {
             carti->push_back((*other.carti)[i]);
         }
     }
